+++ tests/tests_legacy/test_group.py
