+++ tests/tests_legacy/test_join_alias.py
