+++ tests/tests_legacy/test_operators_accumulators.py
