+++ tests/tests_legacy/test_operators_array.py
