+++ tests/tests_legacy/test_operators_boolean.py
