+++ tests/tests_legacy/test_operators_comparison.py
