+++ tests/tests_legacy/test_operators_objects.py
