+++ tests/tests_legacy/test_search_operators.py
