+++ tests/tests_legacy/test_search_pipeline.py
