+++ tests/tests_legacy/test_signatures.py
