+++ tests/tests_legacy/test_stages.py
